+++ services/blockchain-indexer/shared/dataService/index.js
@@ -84,7 +84,6 @@ const {
 	getPendingTransactions,
 	reloadAllPendingTransactions,
 	postTransactions,
-	getSchemas,
 	getTransactionsByBlockID,
 	dryRunTransactions,
 } = require('./transactions');
@@ -108,6 +107,7 @@ const {
 const { getEvents } = require('./events');
 const { getAuthAccountInfo } = require('./auth');
 const { getValidator, validateBLSKey } = require('./validator');
+const { getSchemas } = require('./schemas');
 
 module.exports = {
 	getMultisignatureGroups,
